visualization-agent/requirements.txt
streamlit
google-cloud-bigquery
google-cloud-storage
pandas
pydeck
db-dtypes
rasterio
geopandas
shapely